# frontend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Village API",
    "version": "0.1.0"
  },
  "paths": {
...
    "\/run\/list": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "description": "List all runs.",
        "operationId": "list_runs",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "title": "Response List Runs",
                  "type": "array",
                  "items": {
                    "$ref": "#\/components\/schemas\/RunWithScript"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/run\/get": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "description": "Get a run.",
        "operationId": "get_run",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            },
            "name": "run_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/RunWithScriptDetailed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/run\/delete": {
      "delete": {
        "tags": [
          "runs"
        ],
        "summary": "Delete Run",
        "description": "Delete a run.",
        "operationId": "delete_run",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            },
            "name": "run_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/schedule\/create": {
      "post": {
        "tags": [
          "schedules"
        ],
        "summary": "Create Schedule",
        "description": "Add a script schedule to the scheduler",
        "operationId": "create_schedule",
        "requestBody": {
          "content": {
            "application\/json": {
              "schema": {
                "$ref": "#\/components\/schemas\/CreateScheduleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Schedule"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/schedule\/update": {
      "post": {
        "tags": [
          "schedules"
        ],
        "summary": "Update Schedule",
        "operationId": "update_schedule",
        "requestBody": {
          "content": {
            "application\/json": {
              "schema": {
                "$ref": "#\/components\/schemas\/UpdateScheduleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Schedule"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/schedule\/run": {
      "post": {
        "tags": [
          "schedules"
        ],
        "summary": "Run Schedule",
        "description": "Execute a script based on its schedule entry",
        "operationId": "run_schedule",
        "requestBody": {
          "content": {
            "application\/json": {
              "schema": {
                "$ref": "#\/components\/schemas\/RunScheduleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "\/schedule\/list": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "List Schedules",
        "description": "List all schedules",
        "operationId": "list_schedules",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "title": "Response List Schedules",
                  "type": "array",
                  "items": {
                    "$ref": "#\/components\/schemas\/ScheduleWithScript"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/schedule\/delete": {
      "delete": {
        "tags": [
          "schedules"
        ],
        "summary": "Delete Schedule",
        "description": "Delete a script schedule from the scheduler",
        "operationId": "delete_schedule",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            },
            "name": "schedule_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Schedule"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/schedule\/get": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "Get Schedule",
        "description": "Get a script schedule from the scheduler",
        "operationId": "get_schedule",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            },
            "name": "schedule_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/ScheduleWithMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}